python/backend/requirements.txt
Flask
TensorFlow
Pillow
numpy
CORS
